backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.18
anthropic==0.43.0
docker==7.1.0
httpx==0.28.1
pymupdf==1.25.3
python-dotenv==1.0.1
pydantic==2.10.4
aiofiles==24.1.0
pyyaml==6.0.2
